Update README.md
Changed this in the docs because it was all wrong. We set it to "yes" and nothing worked. No error, no nothing. Subscribing did not make a difference. The field needed to be set to "1", not to "yes". So a code example changing default values is not great. Especially if it sets values to "1" instead of "yes", which makes it even harder to figure out what this is even for.